+++ frontend/src/components/BackgroundMusic/BackgroundMusic.tsx
@@ -1,7 +1,9 @@
-import CustomButton from '../CustomButton';
-import { Icon } from '@iconify/react';
 import axios from 'axios';
 import { useEffect, useState } from 'react';
+
+import { CustomButton } from '@components/Buttons';
+
+import { Icon } from '@iconify/react';
 
 const GET_BGM_API =
   'https://www.viodio.io/_api/v1/musics/627c8e78715664b72b9bee90/audios/dcbfe1206d7b4410a7e71963bc9e2e4d?auto_play=true';
