+++ frontend/src/components/Buttons/index.ts
@@ -1,0 +1,5 @@
+import CustomButton from './CustomButton';
+import IconButton from './IconButton';
+import LogoButton from './LogoButton';
+
+export { CustomButton, IconButton, LogoButton };
