--- frontend/src/components/CustomButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CustomButton';
