--- frontend/src/components/IconButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './IconButton';
