+++ frontend/src/pages/AIChatPage/AIChatPage.tsx
@@ -1,13 +1,12 @@
 import { useRef } from 'react';
 
 import Background from '@components/Background';
+import { CustomButton } from '@components/Buttons';
 import ChatInput from '@components/ChatInput';
 import ChatList from '@components/ChatList';
-import CustomButton from '@components/CustomButton';
 import Header from '@components/Header';
 
 import { useAiChatMessage, useTarotSpread } from '@business/hooks/useAiChat';
-
 
 import { Icon } from '@iconify/react';
