+++ frontend/src/pages/HomePage/HomePage.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from 'react-router-dom';
 
 import Background from '@components/Background';
-import CustomButton from '@components/CustomButton';
+import { CustomButton } from '@components/Buttons';
 
 import { connect } from '@business/services/socket';
